requirements.txt
numpy==1.26.4
numba==0.60.0
Pillow==10.3.0
opencv-python==4.10.0.84

fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
PyQt6==6.7.1

# ML
ultralytics==8.2.45
ncnn==1.0.20240410
pandas~=2.2.2
PyYAML~=6.0.1
scikit-learn~=1.5.2
torch~=2.3.1
transformers~=4.47.1